+++ Commands/Provisioning/Site/ApplyProvisioningTemplate.cs
@@ -161,8 +161,16 @@ PS:> Apply-PnPProvisioningTemplate -Path NewTemplate.xml -ExtensibilityHandlers 
                 XMLTemplateProvider provider;
                 if (isOpenOfficeFile)
                 {
-                    provider = new XMLOpenXMLTemplateProvider(new OpenXMLConnector(templateFileName, fileConnector));
-                    templateFileName = templateFileName.Substring(0, templateFileName.LastIndexOf(".", StringComparison.Ordinal)) + ".xml";
+                    var openXmlConnector = new OpenXMLConnector(templateFileName, fileConnector);
+                    provider = new XMLOpenXMLTemplateProvider(openXmlConnector);
+                    if (!String.IsNullOrEmpty(openXmlConnector.Info?.Properties?.TemplateFileName))
+                    {
+                        templateFileName = openXmlConnector.Info.Properties.TemplateFileName;
+                    }
+                    else
+                    {
+                        templateFileName = templateFileName.Substring(0, templateFileName.LastIndexOf(".", StringComparison.Ordinal)) + ".xml";
+                    }
                 }
                 else
                 {