+++ CHANGELOG.md
@@ -9,10 +9,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Save-PnPClientSidePageConversionLog: use this cmdlet to save the pending page transformation logs. Needs to be used in conjunction with the -LogSkipFlush flag on the ConvertTo-PnPClientSidePage cmdlet.
+
 ### Changed
 
 - Cleanup private key only for file or pem based certificate login (#2101)
 - ConvertTo-PnPClientSidePage: Added support to transform web part pages that live outside of a library (so in the root of the site)
+- ConvertTo-PnPClientSidePage: Added support to specify the target site as a connection using the TargetConnection parameter. This allows to read a page in one environment (on-premises, tenant A) and create in another online location (tenant B). (#2098)
+
+### Contributors
+
+ Paul Bullock [pkbullock]
 
 ## [3.9.1905.3 - May 2019 Intermediate Release 3]